moved results outside of test and each result specific to a quiz
