+++ client/src/types/quiz.ts
@@ -15,6 +15,18 @@ export interface AdminQuiz {
   status: number;
 }
 
+export interface AdminQuizResponse {
+  count: number;
+  next: string;
+  previous: string;
+  results: AdminQuiz[];
+}
+
+export interface AdminQuizNameResponse {
+  id: number;
+  name: string;
+}
+
 export interface AdminQuizSlotRequest {
   question_id: number;
   slot_index: number;
